pyproject.toml
[project]
name = "cliTeXlib"
version = "0.1.0"
description = "A CLI-based TeX project library."
authors = [
    {name = "ziang-alpha",email = "zzalpha2024@163.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "click (>=8.3.1,<9.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "python-frontmatter (>=1.1.0,<2.0.0)",
    "rich (>=14.3.3,<15.0.0)"
]

[project.scripts]
tlib = "main.tlib:tlib"


[tool.poetry]
# ...
packages = [
    { include = "main", from = "src" },
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
